requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
Werkzeug==2.0.2
